Reorganize template_manager for extension
