+++ lib/src/services/config_services/gitignore_manager.dart
