+++ lib/src/services/template_manager.dart
@@ -1,0 +1,3 @@
+class TemplateManager {
+  
+}
